backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pymongo==4.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.106.0
PyYAML==6.0.1
PyPDF2==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
